+++ rioxarray/_io.py
@@ -811,15 +811,15 @@ def open_rasterio(
         filename = vrt.src_dataset.name
         vrt_params = dict(
             src_crs=vrt.src_crs.to_string() if vrt.src_crs else None,
-            crs=vrt.crs.to_string() if vrt.crs else None,
+            crs=vrt.dst_crs.to_string() if vrt.dst_crs else None,
             resampling=vrt.resampling,
             tolerance=vrt.tolerance,
             src_nodata=vrt.src_nodata,
-            nodata=vrt.nodata,
-            width=vrt.width,
-            height=vrt.height,
+            nodata=vrt.dst_nodata,
+            width=vrt.dst_width,
+            height=vrt.dst_height,
             src_transform=vrt.src_transform,
-            transform=vrt.transform,
+            transform=vrt.dst_transform,
             dtype=vrt.working_dtype,
             warp_extras=vrt.warp_extras,
         )